+++ goopylib/src/Platform/Independent/BaseVertexArray.cpp
@@ -1,6 +1,14 @@
 #include "BaseVertexArray.h"
 
 namespace gp {
+    BaseVertexArray::BaseVertexArray() {
+        GP_CORE_INFO("Initializing Vertex Array");
+    }
+
+    BaseVertexArray::~BaseVertexArray() {
+        GP_CORE_INFO("Deallocating Vertex Array");
+    }
+
     const std::vector<std::shared_ptr<BaseVertexBuffer>> &BaseVertexArray::getVertexBuffers() const {
         return m_VertexBuffers;
     }